HotelFinder/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
peewee==3.16.2
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.31.0
urllib3==2.0.3
